+++ superglue/jest.config.js
@@ -1,18 +1,9 @@
-// jest.config.js
-// Sync object
-/** @type {import('@jest/types').Config.InitialOptions} */
-const config = {
-  verbose: true,
-};
-
-module.exports = config;
-
-// Or async function
 module.exports = async () => {
   return {
     modulePathIgnorePatterns: ["<rootDir>/dist/"],
     verbose: true,
     testEnvironment: 'jsdom',
+    clearMocks: true,
     setupFiles: [
       "<rootDir>/spec/helpers/setup.js",
       "<rootDir>/spec/helpers/polyfill.js"