+++ superglue/spec/lib/NavComponent.spec.js
@@ -40,6 +40,8 @@ describe('Nav', () => {
       history.push("/bar", {
         superglue: true,
         pageKey: '/bar',
+        posX: 0,
+        posY: 0
       })
       let expected = '<div><h1>About Page</h1></div>'
 
@@ -49,8 +51,8 @@ describe('Nav', () => {
       const mockStore = configureMockStore()
       const store = mockStore({
         pages: {
-          '/foo': { componentIdentifier: 'about' },
-          '/bar': { componentIdentifier: 'home' },
+          '/foo': { componentIdentifier: 'about', restoreStrategy: 'fromCacheOnly' },
+          '/bar': { componentIdentifier: 'home', restoreStrategy: 'fromCacheOnly' },
         },
         superglue: {
           csrfToken: "abc"
@@ -67,10 +69,12 @@ describe('Nav', () => {
           />
         </Provider>
       )
+      const scrollTo = jest.spyOn(window, 'scrollTo').mockImplementation(() => {});
       expect(component.find(Home).exists()).toBe(true)
       expect(component.find(About).exists()).toBe(false)
 
       component.find('button').simulate('click')
+      expect(scrollTo).toHaveBeenCalledWith(0, 0)
 
       expect(component.find(Home).exists()).toBe(false)
       expect(component.find(About).exists()).toBe(true)
@@ -81,7 +85,7 @@ describe('Nav', () => {
       const mockStore = configureMockStore()
       const store = mockStore({
         pages: {
-          '/bar': { componentIdentifier: 'home' },
+          '/bar': { componentIdentifier: 'home', restoreStrategy: 'fromCacheOnly' },
         },
         superglue: {
           csrfToken: "abc"
@@ -113,18 +117,20 @@ describe('Nav', () => {
       history.push("/bar", {
         superglue: true,
         pageKey: '/bar',
+        posX: 0,
+        posY: 0
       })
       const mockStore = configureMockStore()
       const store = mockStore({
         pages: {
-          '/foo': { componentIdentifier: 'about' },
-          '/bar': { componentIdentifier: 'home' },
+          '/foo': { componentIdentifier: 'about', restoreStrategy: 'fromCacheOnly' },
+          '/bar': { componentIdentifier: 'home', restoreStrategy: 'fromCacheOnly' },
         },
         superglue: {
           csrfToken: "abc"
         }
       })
-
+      const scrollTo = jest.spyOn(window, 'scrollTo').mockImplementation(() => {});
       const component = mount(
         <Provider store={store}>
           <Nav
@@ -136,10 +142,19 @@ describe('Nav', () => {
         </Provider>
       )
       component.find('button').simulate('click')
+      expect(scrollTo).toHaveBeenCalledWith(0, 0)
       expect(component.find(Home).exists()).toBe(false)
       expect(component.find(About).exists()).toBe(true)
 
       const expectedActions = [
+        {
+          type: '@@superglue/HISTORY_CHANGE',
+          payload: {
+            pathname: '/bar',
+            search: '',
+            hash: '',
+          }
+        },
         {
           type: '@@superglue/HISTORY_CHANGE',
           payload: {
@@ -157,12 +172,14 @@ describe('Nav', () => {
       history.push("/bar", {
         superglue: true,
         pageKey: '/bar',
+        posX: 0,
+        posY: 0,
       })
       const mockStore = configureMockStore()
       const store = mockStore({
         pages: {
-          '/bar': { componentIdentifier: 'home' },
-          '/foo': { componentIdentifier: 'about' },
+          '/bar': { componentIdentifier: 'home', restoreStrategy: 'fromCacheOnly' },
+          '/foo': { componentIdentifier: 'about', restoreStrategy: 'fromCacheOnly' },
         },
         superglue: {
           csrfToken: "abc"
@@ -185,7 +202,7 @@ describe('Nav', () => {
       class ExampleAbout extends About {
         componentDidMount() {
           visitedAbout = true
-          history.goBack()
+          history.back()
         }
       }
 
@@ -203,6 +220,245 @@ describe('Nav', () => {
       component.find('button').simulate('click')
       expect(visitedAbout).toBe(true)
       expect(returnedHome).toBe(true)
+    })
+
+    it('returns false when action is none', () => {
+      const history = createMemoryHistory({})
+      const mockStore = configureMockStore()
+      const store = mockStore({
+        pages: {
+          '/bar': { componentIdentifier: 'home', restoreStrategy: 'fromCacheOnly' },
+          '/foo': { componentIdentifier: 'about', restoreStrategy: 'fromCacheOnly' },
+        },
+        superglue: {
+          csrfToken: "abc"
+        }
+      })
+      const component = mount(
+        <Nav
+          store={store}
+          mapping={{ home: Home, about: About }}
+          initialPageKey={'/bar'}
+          history={history}
+        />
+      )
+
+      expect(
+        component.instance().navigateTo('/blah', {action: 'none'})
+      ).toEqual(false)
+    })
+  })
+
+  describe('history pop', () => {
+    describe('when the previous page was set to "revisitOnly"', () => {
+      it('revisits the page and scrolls when finished', () => {
+        const history = createMemoryHistory({})
+        history.replace("/bar", {
+          superglue: true,
+          pageKey: '/bar',
+          posX: 5,
+          posY: 5
+        })
+
+        history.push("/foo", {
+          superglue: true,
+          pageKey: '/foo',
+          posX: 10,
+          posY: 10
+        })
+        const mockStore = configureMockStore()
+        const store = mockStore({
+          pages: {
+            '/bar': { componentIdentifier: 'home', restoreStrategy: 'revisitOnly' },
+            '/foo': { componentIdentifier: 'about', restoreStrategy: 'revisitOnly' },
+          },
+          superglue: {
+            csrfToken: "abc"
+          }
+        })
+        const scrollTo = jest.spyOn(window, 'scrollTo').mockImplementation(() => {});
+        const suggestedAction = 'none'
+
+        const fakeVisit = jest.fn((...args) => {
+          return {
+            then: jest.fn(fn => {
+              expect(scrollTo).not.toHaveBeenCalled()
+              fn({suggestedAction})
+              expect(scrollTo).toHaveBeenCalledWith(5, 5)
+            })
+          }
+        })
+
+        const component = mount(
+          <Provider store={store}>
+            <Nav
+              store={store}
+              visit={fakeVisit}
+              mapping={{ home: Home, about: About }}
+              initialPageKey={'/bar'}
+              history={history}
+            />
+          </Provider>
+        )
+
+        history.back()
+        expect(fakeVisit).toHaveBeenCalledWith("/bar", { revisit: true })
+      })
+
+      it('revisits the page and skips scroll when redirected (suggestedAction is not "none")', () => {
+        const history = createMemoryHistory({})
+        history.replace("/bar", {
+          superglue: true,
+          pageKey: '/bar',
+          posX: 5,
+          posY: 5
+        })
+
+        history.push("/foo", {
+          superglue: true,
+          pageKey: '/foo',
+          posX: 10,
+          posY: 10
+        })
+        const mockStore = configureMockStore()
+        const store = mockStore({
+          pages: {
+            '/bar': { componentIdentifier: 'home', restoreStrategy: 'revisitOnly' },
+            '/foo': { componentIdentifier: 'about', restoreStrategy: 'revisitOnly' },
+          },
+          superglue: {
+            csrfToken: "abc"
+          }
+        })
+        const scrollTo = jest.spyOn(window, 'scrollTo').mockImplementation(() => {});
+        const suggestedAction = 'push'
+
+        const fakeVisit = jest.fn((...args) => {
+          return {
+            then: jest.fn(fn => {
+              // expect(scrollTo).not.toHaveBeenCalled()
+              fn({suggestedAction})
+              expect(scrollTo).not.toHaveBeenCalled()
+            })
+          }
+        })
+
+        // scroll to 0 0
+
+        const component = mount(
+          <Provider store={store}>
+            <Nav
+              store={store}
+              visit={fakeVisit}
+              mapping={{ home: Home, about: About }}
+              initialPageKey={'/bar'}
+              history={history}
+            />
+          </Provider>
+        )
+
+        history.back()
+        expect(fakeVisit).toHaveBeenCalledWith("/bar", { revisit: true })
+      })
+    })
+
+    describe('when the previous page was set to "fromCacheOnly"', () => {
+      it('restores without visiting and scrolls', () => {
+        const history = createMemoryHistory({})
+        history.replace("/bar", {
+          superglue: true,
+          pageKey: '/bar',
+          posX: 5,
+          posY: 5
+        })
+
+        history.push("/foo", {
+          superglue: true,
+          pageKey: '/foo',
+          posX: 10,
+          posY: 10
+        })
+        const mockStore = configureMockStore()
+        const store = mockStore({
+          pages: {
+            '/bar': { componentIdentifier: 'home', restoreStrategy: 'fromCacheOnly' },
+            '/foo': { componentIdentifier: 'about', restoreStrategy: 'fromCacheOnly' },
+          },
+          superglue: {
+            csrfToken: "abc"
+          }
+        })
+        const scrollTo = jest.spyOn(window, 'scrollTo').mockImplementation(() => {});
+        const suggestedAction = 'none'
+
+        const fakeVisit = jest.fn()
+
+        const component = mount(
+          <Provider store={store}>
+            <Nav
+              store={store}
+              visit={fakeVisit}
+              mapping={{ home: Home, about: About }}
+              initialPageKey={'/bar'}
+              history={history}
+            />
+          </Provider>
+        )
+
+        history.back()
+        expect(scrollTo).toHaveBeenCalledWith(5, 5)
+        expect(fakeVisit).not.toHaveBeenCalled()
+      })
+    })
+
+    describe('and the previous page was set to "fromCacheAndRevisitInBackground"', () => {
+      it('restores, scrolls, and revisits the page in the background', () => {
+        const history = createMemoryHistory({})
+        history.replace("/bar", {
+          superglue: true,
+          pageKey: '/bar',
+          posX: 5,
+          posY: 5
+        })
+
+        history.push("/foo", {
+          superglue: true,
+          pageKey: '/foo',
+          posX: 10,
+          posY: 10
+        })
+        const mockStore = configureMockStore()
+        const store = mockStore({
+          pages: {
+            '/bar': { componentIdentifier: 'home', restoreStrategy: 'fromCacheAndRevisitInBackground' },
+            '/foo': { componentIdentifier: 'about', restoreStrategy: 'fromCacheAndRevisitInBackground' },
+          },
+          superglue: {
+            csrfToken: "abc"
+          }
+        })
+        const scrollTo = jest.spyOn(window, 'scrollTo').mockImplementation(() => {});
+        const suggestedAction = 'none'
+
+        const fakeVisit = jest.fn((...args) => {
+          expect(scrollTo).toHaveBeenCalledWith(5, 5)
+        })
+
+        const component = mount(
+          <Provider store={store}>
+            <Nav
+              store={store}
+              visit={fakeVisit}
+              mapping={{ home: Home, about: About }}
+              initialPageKey={'/bar'}
+              history={history}
+            />
+          </Provider>
+        )
+
+        history.back()
+        expect(fakeVisit).toHaveBeenCalledWith("/bar", { revisit: true })
+      })
     })
   })
 })